backend/requirements.txt
fastapi==0.118.0
uvicorn[standard]==0.35.0
sqlalchemy==2.0.36
pydantic-settings==2.6.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.12
psycopg2-binary==2.9.10
openai==1.60.2
pytest==8.4.2
httpx==0.28.1
pytest-asyncio==0.24.0
